requirements.txt
flask
qdrant-client
sentence-transformers==2.2.2
huggingface_hub==0.14.1
PyPDF2
python-docx
pandas
google.generativeai
langgraph
Flask-CORS
flask_pymongo
langchain
pytesseract 
pillow
deep-translator
